fix bash check for platform and build_type

+++ ci/github-actions/setup.sh
@@ -2,7 +2,7 @@
 
 echo "The build architecture is ${ImageOS}"
 
-if [ "${ImageOS}" == "docker" ]; then
+if [ "${ImageOS}" == "ubuntu18" ] && [ "${BUILD_TYPE}" == "docker" ]; then
     echo "Installing docker compose"
     sudo rm /usr/local/bin/docker-compose
     curl -L https://github.com/docker/compose/releases/download/${DOCKER_COMPOSE_VERSION}/docker-compose-`uname -s`-`uname -m` > docker-compose
